+++ CesiumGltf/test/TestPropertyAttributeView.cpp
@@ -3,9 +3,9 @@
 #include <CesiumUtility/Assert.h>
 
 #include <catch2/catch.hpp>
-#include <gsl/span>
 
 #include <cstddef>
+#include <span>
 #include <vector>
 
 using namespace CesiumGltf;