+++ CHANGES.md
@@ -2,10 +2,16 @@
 
 ### Not Released Yet
 
+##### Breaking Changes :mega:
+
+- Cesium Native now requires C++20.
+- Switched from `gsl::span` to `std::span` throughout the library and API. The GSL library has been removed.
+
 ##### Additions :tada:
 
 - Added support for `EXT_accessor_additional_types` in `AccessorView`.
 - Added `EllipsoidTilesetLoader` that will generate a tileset by tesselating the surface of an ellipsoid, producing a simple globe tileset without any terrain features.
+- The `schemaUri` property in the `EXT_structural_metadata` glTF extension is now supported, allowing structural metadata schemas to be loaded from URIs rather than being embedded in the glTF itself.
 
 ##### Fixes :wrench: